requirements.txt
ipykernel
numpy==1.26.4
ultralytics==8.3.12
torchvision
from_root
Pillow
pyyaml
uuid
flask
flask_limiter
-e .